backend/requirements.txt
# FastAPI 및 서버
fastapi==0.117.1
uvicorn[standard]==0.37.0
python-multipart==0.0.20

# 머신러닝/딥러닝
torch==2.7.1
torchvision==0.22.1
transformers==4.45.0
sentence-transformers==3.0.1

# 벡터 검색 (numpy 1.x 필요)
faiss-cpu==1.8.0
llama-cpp-python==0.2.90

# 데이터 처리 (numpy 1.x 호환)
numpy==1.26.4
pandas==2.3.2
scikit-learn==1.7.1

# 기타 유틸
python-dotenv==1.0.0
pydantic==2.11.9
huggingface-hub>=0.23.2

# TripPrep
anthropic
tavily-python
